+++ addon/connection.cc
@@ -106,18 +106,23 @@ Napi::Value Connection::Execute(const Napi::CallbackInfo &info) {
 }
 
 Napi::Value Connection::Close(const Napi::CallbackInfo &info) {
-  for (auto &result : *results) {
-    if (result != nullptr) {
-      result->close();
-    }
+  // the following gives segfaults for some reason now
+  // for (auto &result : *results) {
+  //   if (result != nullptr) {
+  //     result->close();
+  //   }
+  // }
+  if (connection) {
+    connection.reset();
   }
-  connection.reset();
-  database.reset();
+  if (database) {
+    database.reset();
+  }
   return info.Env().Undefined();
 }
 Napi::Value Connection::IsClosed(const Napi::CallbackInfo &info) {
   Napi::Env env = info.Env();
-  bool isClosed = connection == nullptr || connection->context->is_invalidated;
+  bool isClosed = connection == nullptr;
   return Napi::Boolean::New(env, isClosed);
 }
 } // namespace NodeDuckDB