+++ addon/duckdb.h
@@ -9,7 +9,7 @@ class DuckDB : public Napi::ObjectWrap<DuckDB> {
 public:
   static Napi::Object Init(Napi::Env env, Napi::Object exports);
   DuckDB(const Napi::CallbackInfo &info);
-  shared_ptr<duckdb::DuckDB> database;
+  duckdb::shared_ptr<duckdb::DuckDB> database;
   static Napi::FunctionReference constructor;
   bool IsClosed(void);