+++ addon/result_iterator.h
@@ -12,7 +12,7 @@ public:
   static Napi::Object Init(Napi::Env env, Napi::Object exports);
   ResultIterator(const Napi::CallbackInfo &info);
   static Napi::Object Create();
-  unique_ptr<duckdb::QueryResult> result;
+  duckdb::unique_ptr<duckdb::QueryResult> result;
   ResultFormat rowResultFormat;
   void close();
 
@@ -23,7 +23,7 @@ private:
   Napi::Value GetType(const Napi::CallbackInfo &info);
   Napi::Value Close(const Napi::CallbackInfo &info);
   Napi::Value IsClosed(const Napi::CallbackInfo &info);
-  unique_ptr<duckdb::DataChunk> current_chunk;
+  duckdb::unique_ptr<duckdb::DataChunk> current_chunk;
   uint64_t chunk_offset = 0;
   Napi::Value getCellValue(Napi::Env env, duckdb::idx_t col_idx);
   Napi::Value getRowArray(Napi::Env env);