+++ CHANGELOG.md
@@ -2,6 +2,20 @@
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
+### [0.0.64](https://github.com/deepcrawl/node-duckdb/compare/v0.0.61...v0.0.64) (2021-03-26)
+
+
+### Features
+
+* duckdb v  6fe573 ([#74](https://github.com/deepcrawl/node-duckdb/issues/74)) ([3b5ceca](https://github.com/deepcrawl/node-duckdb/commit/3b5cecad3fc5b692ca13c35cb4b76d5440604810))
+
+### [0.0.61](https://github.com/deepcrawl/node-duckdb/compare/v0.0.59...v0.0.61) (2021-02-03)
+
+
+### Features
+
+* duckdb v  6fe573 ([#74](https://github.com/deepcrawl/node-duckdb/issues/74)) ([3b5ceca](https://github.com/deepcrawl/node-duckdb/commit/3b5cecad3fc5b692ca13c35cb4b76d5440604810))
+
 ### [0.0.59](https://github.com/deepcrawl/node-duckdb/compare/v0.0.58...v0.0.59) (2020-12-16)
 
 ### [0.0.58](https://github.com/deepcrawl/node-duckdb/compare/v0.0.57...v0.0.58) (2020-12-16)